+++ redpen-core/src/main/resources/cc/redpen/validator/document/error-messages.properties
@@ -2,3 +2,4 @@ UnexpandedAcronymValidator.UnexpandedAcronym=The expanded form of the acronym {0
 FrequentSentenceStartValidator.SentenceStartTooFrequent={0}% of sentences start with "{1}".
 WordFrequencyValidator.WordUsedTooFrequently=The word "{0}" comprises {1}% of all words. Expected around {2}%.
 GappedSectionValidator=The section "{0}" (level {1}) is misplaced. It should have level of {2}.
+JapaneseExpressionVariationValidator=Found possible Japanese word variations: "{0}"  and "{1}"
