+++ redpen-core/src/main/resources/cc/redpen/validator/document/error-messages_ja.properties
@@ -2,3 +2,4 @@ UnexpandedAcronymValidator.UnexpandedAcronym=\u7701\u7565 "{0}"\u3001\u306E\u6B6
 FrequentSentenceStartValidator.SentenceStartTooFrequent={0}% \u306E\u6587\u304C "{1}" \u304B\u3089\u306F\u3058\u307E\u3063\u3066\u3044\u307E\u3059\u3002
 WordFrequencyValidator.WordUsedTooFrequently=\u5358\u8A9E "{0}" \u306E\u5229\u7528\u7387\u304C {1}% \u3067\u3059\u3002\u3053\u306E\u5358\u8A9E\u306E\u4F7F\u7528\u7387\u306F "{2}" \u7A0B\u5EA6\u304C\u671B\u307E\u3057\u3044\u3067\u3059\u3002
 GappedSectionValidator=\u7BC0 "{0}" \uFF08\u30EC\u30D9\u30EB {1}\uFF09\u306E\u914D\u7F6E\u304C\u6B63\u3057\u304F\u3042\u308A\u307E\u305B\u3093\u3002\u30EC\u30D9\u30EB {2} \u3067\u3042\u308B\u5FC5\u8981\u304C\u3042\u308A\u307E\u3059\u3002
+JapaneseExpressionVariationValidator=\u5358\u8A9E\u306E\u63FA\u3089\u304E\u306E\u53EF\u80FD\u6027\u304C\u3042\u308B\u8868\u73FE\u3001 \u201D{0}\u201D \u3001\u201D{1}\u201D\u304C\u898B\u3064\u304B\u308A\u307E\u3057\u305F\u3002
