+++ redpen-core/src/main/resources/cc/redpen/validator/sentence/error-messages.properties
@@ -48,4 +48,4 @@ JapaneseBrokenExpressionValidator=Found broken japanese expression (missing "Ra"
 JapaneseJoyoKanjiValidator=Found non-joyo kanji: "{0}"
 DoubledConjunctiveParticleGaValidator=Found multiple conjunctive particle: "{0}"
 SuccessiveSentenceValidator=Found similar two sentences in succession: "{0}" and "{1}"
-JapaneseExpressionVariationValidator=Found possible Japanese word variations: "{0}"  and "{1}"
+
