+++ redpen-core/src/main/resources/cc/redpen/validator/sentence/error-messages_ja.properties
@@ -48,4 +48,4 @@ JapaneseBrokenExpressionValidator=\u3089\u629C\u304D\u8868\u73FE\u3092\u4F7F\u75
 JapaneseJoyoKanjiValidator=\u5E38\u7528\u6F22\u5B57\u3067\u306A\u3044\u6F22\u5B57\uFF08"{0}"\uFF09\u304C\u4F7F\u7528\u3055\u308C\u3066\u3044\u307E\u3059\u3002
 DoubledConjunctiveParticleGaValidator=\u4E00\u6587\u306B\u9006\u8AAC\u306E\u63A5\u7D9A\u52A9\u8A5E "{0}" \u304C\u8907\u6570\u56DE\u4F7F\u7528\u3055\u308C\u3066\u3044\u307E\u3059\u3002
 SuccessiveSentenceValidator=\u985E\u4F3C\u3059\u308B\u6587\u304C\u4E8C\u3064\u7D9A\u3051\u3066\u4F7F\u7528\u3055\u308C\u3066\u3044\u307E\u3059: "{0}"\u3001"{1}"
-JapaneseExpressionVariationValidator=\u5358\u8A9E\u306E\u63FA\u3089\u304E\u306E\u53EF\u80FD\u6027\u304C\u3042\u308B\u8868\u73FE\u3001 \u201D{0}\u201D \u3001\u201D{1}\u201D\u304C\u898B\u3064\u304B\u308A\u307E\u3057\u305F\u3002
+
